renamed: substratee-node-calls -> substratee-node-primitives: put all the type definitions of the node inside there in order to have them at one place.
