+++ enclave/Cargo.toml
@@ -117,5 +117,10 @@ features = ["sgx"]
 git = "https://github.com/scs/sgx-runtime"
 default-features = false
 
+[dependencies.substratee-node-primitives]
+path = "../substratee-node-primitives"
+default-features = false
+features = ["sgx"]
+
 [patch.crates-io]
 sp-io = { git = "https://github.com/scs/sgx-runtime", default-features = false, features = ["disable_oom", "disable_panic_handler", "disable_allocator", "sgx"]}
