--- substratee-node-calls/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "substratee-node-calls"
-version = "0.6.1-sub2.0.0-alpha.7"
-authors = ["clangenbacher <christian.langenbacher@scs.ch>"]
-edition = "2018"
-
-[dependencies]
-log = "0.4"
-codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false, features = ["derive"] }
-primitive-types = { version = "0.6", default-features = false, features = ["codec"] }
-base58 			        = "0.1"
-
-[dependencies.my_node_runtime]
-git = "https://github.com/scs/substraTEE-node"
-tag = "v0.6.3-sub2.0.0-alpha.7"
-package = "substratee-node-runtime"
-
-[dependencies.substrate-api-client]
-git = "https://github.com/scs/substrate-api-client"
-#tag = "v0.4.3-sub2.0.0-alpha.7"
-branch = "cl-make-url-public"
-
-[dependencies.sp-core]
-version = '2.0.0-alpha.7'
-
-[dependencies.sp-runtime]
-version = '2.0.0-alpha.7'
