+++ substratee-node-primitives/Cargo.toml
@@ -1,0 +1,49 @@
+[package]
+name = "substratee-node-primitives"
+version = "0.6.1-sub2.0.0-alpha.7"
+authors = ["clangenbacher <christian.langenbacher@scs.ch>"]
+edition = "2018"
+
+[dependencies]
+log             = { version = "0.4", optional = true }
+codec           = { package = "parity-scale-codec", version = "1.3.0", default-features = false, features = ["derive"] }
+primitive-types = { version = "0.6", default-features = false, features = ["codec"] }
+base58 			= { version = "0.1", optional = true }
+
+[dependencies.sgx_tstd]
+git = "https://github.com/apache/teaclave-sgx-sdk.git"
+features = ["untrusted_fs","net","backtrace"]
+rev = "v1.1.2"
+optional = true
+
+[dependencies.my-node-runtime]
+git = "https://github.com/scs/substraTEE-node"
+tag = "v0.6.3-sub2.0.0-alpha.7"
+package = "substratee-node-runtime"
+optional = true
+default-features = false
+
+[dependencies.substrate-api-client]
+git = "https://github.com/scs/substrate-api-client"
+#tag = "v0.4.3-sub2.0.0-alpha.7"
+branch = "cl-make-url-public"
+optional = true
+
+[dependencies.sp-core]
+version = '2.0.0-alpha.7'
+default-features = false
+
+[dependencies.sp-runtime]
+version = '2.0.0-alpha.7'
+default-features = false
+optional = true
+
+[features]
+default = ['std']
+std = [ 'substrate-api-client',
+        'my-node-runtime',
+        'sp-core/std',
+        'sp-runtime',
+        'base58',
+        'log']
+sgx = [ 'sgx_tstd' ]
